% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/agg_allGenerics.R, R/aggregate_by_STF.R
\docType{methods}
\name{aggregate}
\alias{aggregate}
\alias{aggregate,Track}
\alias{aggregate,Track-method}
\alias{aggregate,Tracks}
\alias{aggregate,Tracks-method}
\alias{aggregate,TracksCollection}
\alias{aggregate,TracksCollection-method}
\title{Spatio-temporal aggregation of trajectory data}
\usage{
\S4method{aggregate}{Track}(x, by, FUN = mean, ..., simplify = TRUE,
  use.data = TRUE, weight.points = NULL)

\S4method{aggregate}{Tracks}(x, by, FUN = mean, ..., simplify = TRUE,
  use.data = TRUE, weight.points = NULL, weight.tracks = NULL)

\S4method{aggregate}{TracksCollection}(x, by, FUN = mean, ...,
  simplify = TRUE, use.data = TRUE, weight.points = NULL,
  weight.tracks = NULL, byID = FALSE)
}
\arguments{
\item{x}{object of class \code{Track}, \code{Tracks} or \code{TracksCollection}
(see section 'Usage')}

\item{by}{object of class \code{SpatialGrid}, \code{SpatialPixels}, \code{SpatialPolyons}
or \code{STF} with \code{SpatialPixels} or \code{SpatialPolyons} in the \code{sp} slot. \code{data.frame}
counterparts of these objects are accepted as well.}

\item{FUN}{aggregation function used to calculate the aggregated values, optional a function that expects
a vector with weights as its second argument to perform a weighted aggregation.}

\item{...}{optional additional arguments to \code{FUN}}

\item{simplify}{logical, should the return object be simplified (if possible)}

\item{use.data}{TRUE (all data), an index vector or a vector of column names, indicating which
data should be considered from the data slot(s) of the trajectories object.}

\item{weight.points}{\code{NULL}, 'byTime', 'byDist' or 'equal', indicates the weighting of
points' attribute data when calculating the aggregated values from a \code{Track} object's
attributes for a spatio-temporal geometry.}

\item{weight.tracks}{\code{NULL}, 'byTime', 'byDist' or 'equal', indicates the weighting of
track objects when calculating the aggregated values from several \code{Track} objects' parts
for a spatio-temporal geometry.}

\item{byID}{logical, indicating if the aggregation should be performed individually for each
\code{Tracks} object if \code{x} is a \code{TracksCollection}}
}
\value{
An object of class \code{STFDF} is returned, whose data slot contains the aggregated
values for each spatio-temporal geometry.
}
\description{
The method \code{aggregate} performs spatio-temporal (optionally weighted) aggregation of
 trajectory data over spatial or spatio-temporal geometries.

For further detailed information please see the vignette!
}
\note{
The arguments \code{weight.points} and \code{weight.tracks} should just be specified,
if a function is passed to \code{FUN} that expects a vector with weights!
}
\examples{
## load toy example data
load(system.file("extdata","trajaggr_TestData.RData", package = "trajaggr"), verbose = FALSE)
## Aggregation over STF objects
# Aggregate Track over STF with two time intervals
# FUN = max
agg <- aggregate(Track_A1, stf_Polys_2t, FUN = max)
class(agg)
dim(stf_Polys_2t)
dim(agg)
slot(agg, "data")
# FUN = sum
agg_0 <- aggregate(Track_A1, stf_Polys_2t, FUN = sum, na.rm= TRUE)
slot(agg_0, "data")
# FUN = weighted.mean, weighted by time
agg_1 <- aggregate(Track_A1, stf_Polys_2t, FUN = weighted.mean,
weight.points = "byTime")
slot(agg_1, "data")
##
# Aggregate Tracks over STF with two time intervals
agg_2 <- aggregate(Tracks_A, stf_Polys_2t, FUN = weighted.mean, na.rm = TRUE,
weight.points = "byTime")
slot(agg_2, "data")
##
# Aggregate TracksCollection over STF with two time intervals
agg_3 <- aggregate(TrColl, stf_Polys_2t, FUN = weighted.mean, na.rm = TRUE,
weight.points = "byTime")
slot(agg_3, "data")
##
# Aggregate TracksCollection seperated by individuals over STF with two time
# intervals,
agg_3 <- aggregate(TrColl, stf_Polys_2t, FUN = weighted.mean, na.rm = TRUE,
weight.points = "byTime", byID = TRUE)
slot(agg_3, "data")
}

