% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coerce_allGenerics.R
\docType{methods}
\name{as.ltraj}
\alias{as.ltraj}
\alias{as.ltraj,Track}
\alias{as.ltraj,Track-method}
\alias{as.ltraj,Tracks}
\alias{as.ltraj,Tracks-method}
\alias{as.ltraj,TracksCollection}
\alias{as.ltraj,TracksCollection-method}
\title{Coercion to \code{ltraj} objects}
\usage{
\S4method{as.ltraj}{Track}(obj)

\S4method{as.ltraj}{Tracks}(obj)

\S4method{as.ltraj}{TracksCollection}(obj)
}
\arguments{
\item{obj}{object of class \code{Track}, \code{Tracks} or \code{TracksCollection}}
}
\value{
Object of class \code{ltraj}
}
\description{
Coercion of \code{Track}, \code{Tracks} and \code{TracksCollection} objects
to objects of class \code{ltraj}.

For further details please see the vignette.
}
\note{
The method may be used in two ways: as(obj, "ltraj") or as.ltraj(obj).
}
\examples{
## Load example toy data
load(system.file("extdata","trajaggr_TestData.RData", package = "trajaggr"), verbose = FALSE)
## Example 1: coerce Track to ltraj
slot(Track_A1, "data")
slot(Track_A1, "time")
ltraj_Tr_A1 <- as(Track_A1, "ltraj")
class(ltraj_Tr_A1)
require(adehabitatLT)
summary(ltraj_Tr_A1)
ltraj_Tr_A1[[1]]
attr(ltraj_Tr_A1[[1]], "infolocs")
##
## Example 2: coerce TracksCollection to ltraj
slot(TrColl, "tracksCollectionData")
ltraj_TrColl <- as(TrColl, "ltraj")
class(ltraj_TrColl)
require(adehabitatLT)
summary(ltraj_TrColl)
slot(TrColl[2][2], "data")
attr(ltraj_TrColl[[4]], "infolocs")
}

