% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coerce_allGenerics.R
\docType{methods}
\name{as.Move}
\alias{as.Move}
\alias{as.Move,Track}
\alias{as.Move,Track-method}
\title{Coercion to \code{Move} objects}
\usage{
\S4method{as.Move}{Track}(obj)
}
\arguments{
\item{obj}{object of class \code{Track}}
}
\value{
Object of class \code{Move}
}
\description{
Coercion of \code{Track} objects to objects of class \code{Move}.

For further details please see the vignette!
}
\note{
The method may be used in two ways: as(obj, "Move") or as.Move(obj).
}
\examples{
## Load example toy data
load(system.file("extdata","trajaggr_TestData.RData", package = "trajaggr"), verbose = FALSE)
move_Tr_A1 <- as.Move(Track_A1)
slot(Track_A1, "data")
class(move_Tr_A1)
slot(move_Tr_A1, "data")
slot(Track_A1, "time")
slot(move_Tr_A1, "timestamps")
}

