% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coerce_allGenerics.R
\docType{methods}
\name{as.MoveBurst}
\alias{as.MoveBurst}
\alias{as.MoveBurst,Tracks}
\alias{as.MoveBurst,Tracks-method}
\title{Coercion to \code{MoveBurst} objects}
\usage{
\S4method{as.MoveBurst}{Tracks}(obj)
}
\arguments{
\item{obj}{object of class \code{Tracks}}
}
\value{
Object of class \code{MoveBurst}
}
\description{
Coercion of \code{Tracks} objects to objects of class \code{MoveBurst}.

For further details please see the vignette!
}
\note{
The method may be used in two ways: as(obj, "MoveBurst") or as.MoveBurst(obj).

The coercion to \code{MoveBurst} objects just works if the package \pkg{move}
or at least the package \pkg{sp} is explicitly loaded.
}
\examples{
## Load example toy data
load(system.file("extdata","trajaggr_TestData.RData", package = "trajaggr"), verbose = FALSE)
## Coerce Tracks to MoveBurst
require(move)
moveB_Trcs_A <- as.MoveBurst(Tracks_A)
class(moveB_Trcs_A)
slot(moveB_Trcs_A, "data")
slot(Tracks_A[1], "data")
slot(Tracks_A[2], "data")
}

