% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coerce_allGenerics.R
\docType{methods}
\name{as.MoveStack}
\alias{as.MoveStack}
\alias{as.MoveStack,Tracks}
\alias{as.MoveStack,Tracks-method}
\alias{as.MoveStack,TracksCollection}
\alias{as.MoveStack,TracksCollection-method}
\title{Coercion to \code{MoveStack} objects}
\usage{
\S4method{as.MoveStack}{TracksCollection}(obj)

\S4method{as.MoveStack}{Tracks}(obj)
}
\arguments{
\item{obj}{object of class \code{Tracks} or \code{TracksCollection}}
}
\value{
Object of class \code{MoveStack}
}
\description{
Coercion of \code{Tracks} and \code{TracksCollection} objects
to objects of class \code{MoveStack}.

For further details please see the vignette!
}
\note{
The method may be used in two ways: as(obj, "MoveStack") or as.MoveStack(obj).

The coercion to \code{MoveStack} objects just works if the package \pkg{move}
or at least the package \pkg{sp} is explicitly loaded.
}
\examples{
## Load example toy data
load(system.file("extdata","trajaggr_TestData.RData", package = "trajaggr"), verbose = FALSE)
## Example 1: Coerce Tracks to MoveStack
require(move)
move_Trcs_A <- as.MoveStack(Tracks_A)
slot(Tracks_A[1], "data")
class(move_Trcs_A)
slot(move_Trcs_A, "data")[1:6, ]
slot(Tracks_A[2], "time")
slot(move_Trcs_A, "timestamps")[7:11]
##
## Example 2: Coerce TracksCollection to MoveStack
move_TrColl <- as.MoveStack(TrColl)
class(move_TrColl)
dim(TrColl)
length(unique(slot(move_TrColl, "trackId")))
slot(TrColl[2][2], "data")
slot(move_TrColl, "data")[18:21, 1, drop = FALSE]
}

