% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coerce_allGenerics.R
\docType{methods}
\name{as.Track}
\alias{as.Track}
\alias{as.Track,Move}
\alias{as.Track,Move-method}
\alias{as.Track,MoveBurst}
\alias{as.Track,MoveBurst-method}
\alias{as.Track,ltraj}
\alias{as.Track,ltraj-method}
\title{Coercion to \code{Track} objects}
\usage{
\S4method{as.Track}{Move}(obj)

\S4method{as.Track}{MoveBurst}(obj)

\S4method{as.Track}{ltraj}(obj)
}
\arguments{
\item{obj}{object of class \code{Move}, \code{MoveBurst} or \code{ltraj}}
}
\value{
Object of class \code{Track}
}
\description{
Coercion of \code{Move}, \code{MoveBurst} and \code{ltraj} objects
to objects of class \code{Track}.

Just \code{ltraj} objects that contain exactly one trajectory may be coerced to \code{Track} objects.

For further details please see the vignette!
}
\note{
The method may be used in two ways: as(obj, "Track") or as.Track(obj)
}
\examples{
# load example data
data(wildboars_4Ind_ltraj)
# check object
class(wildboars_4Ind_ltraj)
# Subset a Track, adehabitatLT required
require(adehabitatLT)
oneLtraj <- wildboars_4Ind_ltraj[1]
class(oneLtraj)
length(oneLtraj)
summary(oneLtraj[[1]])
# Coerce to Track
track <- as(oneLtraj, "Track")
class(track)
slot(track, "data")
slot(track, "connections")
}

