% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coerce_allGenerics.R
\docType{methods}
\name{as.Tracks}
\alias{as.Tracks}
\alias{as.Tracks,MoveBurst}
\alias{as.Tracks,MoveBurst-method}
\alias{as.Tracks,MoveStack}
\alias{as.Tracks,MoveStack-method}
\alias{as.Tracks,ltraj}
\alias{as.Tracks,ltraj-method}
\title{Coercion to \code{Tracks} objects}
\usage{
\S4method{as.Tracks}{MoveBurst}(obj)

\S4method{as.Tracks}{MoveStack}(obj)

\S4method{as.Tracks}{ltraj}(obj)
}
\arguments{
\item{obj}{object of class \code{MoveBurst}, \code{MoveStack} or \code{ltraj}}
}
\value{
Object of class \code{Tracks}
}
\description{
Coercion of \code{MoveBurst}, \code{MoveStack} and \code{ltraj} objects
to objects of class \code{Tracks}.

Just \code{ltraj} objects that contain more than one trajectory of one individual may be coerced to \code{Tracks} objects.

For further details please see the vignette!
}
\note{
The method may be used in two ways: as(obj, "Tracks") or as.Tracks(obj).
}
\examples{
## load pigeons example data
data(pigeon_R_moveStack)
class(pigeon_R_moveStack)
unique(slot(pigeon_R_moveStack, "trackId"))
length(pigeon_R_moveStack)
# coerce to Tracks
pigeon_R_Trcs <- as.Tracks(pigeon_R_moveStack)
class(pigeon_R_Trcs)
dim(pigeon_R_Trcs)
names(slot(pigeon_R_Trcs, "tracks"))
}

