% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coerce_allGenerics.R
\docType{methods}
\name{as.TracksCollection}
\alias{as.TracksCollection}
\alias{as.TracksCollection,MoveStack}
\alias{as.TracksCollection,MoveStack-method}
\alias{as.TracksCollection,ltraj}
\alias{as.TracksCollection,ltraj-method}
\title{Coercion to \code{TracksCollection} objects}
\usage{
\S4method{as.TracksCollection}{MoveStack}(obj)

\S4method{as.TracksCollection}{ltraj}(obj)
}
\arguments{
\item{obj}{object of class \code{MoveStack} or \code{ltraj}}
}
\value{
Object of class \code{TracksCollection}
}
\description{
Coercion of \code{MoveStack} and \code{ltraj} objects
to objects of class \code{TracksCollection}.

Just \code{ltraj} objects that contain trajectories from more than one individual
may be coerced to \code{TracksCollection} objects.

For further details please see the vignette!
}
\note{
The method may be used in two ways: as(obj, "TracksCollection") or
as.TracksCollection(obj).
}
\examples{
# load wild boars example data
data(wildboars_4Ind_ltraj)
# check object
class(wildboars_4Ind_ltraj)
length(wildboars_4Ind_ltraj)
require(adehabitatLT)
summary(wildboars_4Ind_ltraj)
# Coerce to TracksCollection
tracksColl <- as(wildboars_4Ind_ltraj, "TracksCollection")
class(tracksColl)
dim(tracksColl)
}

