% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/agg_allGenerics.R, R/count.R
\docType{methods}
\name{count}
\alias{count}
\alias{count,Track,STF}
\alias{count,Track,STF-method}
\alias{count,Track,SpatialGrid}
\alias{count,Track,SpatialGrid-method}
\alias{count,Track,SpatialPixels}
\alias{count,Track,SpatialPixels-method}
\alias{count,Track,SpatialPolygons}
\alias{count,Track,SpatialPolygons-method}
\alias{count,Tracks,STF}
\alias{count,Tracks,STF-method}
\alias{count,Tracks,SpatialGrid}
\alias{count,Tracks,SpatialGrid-method}
\alias{count,Tracks,SpatialPixels}
\alias{count,Tracks,SpatialPixels-method}
\alias{count,Tracks,SpatialPolygons}
\alias{count,Tracks,SpatialPolygons-method}
\alias{count,TracksCollection,STF}
\alias{count,TracksCollection,STF-method}
\alias{count,TracksCollection,SpatialGrid}
\alias{count,TracksCollection,SpatialGrid-method}
\alias{count,TracksCollection,SpatialPixels}
\alias{count,TracksCollection,SpatialPixels-method}
\alias{count,TracksCollection,SpatialPolygons}
\alias{count,TracksCollection,SpatialPolygons-method}
\title{Counting trajectories over spatial or spatio-temporal geometries}
\usage{
\S4method{count}{Track,STF}(x, by)

\S4method{count}{Track,SpatialPolygons}(x, by)

\S4method{count}{Track,SpatialPixels}(x, by)

\S4method{count}{Track,SpatialGrid}(x, by)

\S4method{count}{Tracks,STF}(x, by)

\S4method{count}{Tracks,SpatialPolygons}(x, by)

\S4method{count}{Tracks,SpatialPixels}(x, by)

\S4method{count}{Tracks,SpatialGrid}(x, by)

\S4method{count}{TracksCollection,STF}(x, by, byID = FALSE)

\S4method{count}{TracksCollection,SpatialPolygons}(x, by, byID = FALSE)

\S4method{count}{TracksCollection,SpatialPixels}(x, by, byID = FALSE)

\S4method{count}{TracksCollection,SpatialGrid}(x, by, byID = FALSE)
}
\arguments{
\item{x}{an object of class \code{Track}, \code{Tracks} or \code{TracksCollection}}

\item{by}{an object of class \code{Spatial} or \code{STF} (see section 'Usage')}

\item{byID}{logical; indicating if the counting should be performed individually for each
\code{Tracks} object (in the case that a \code{TracksCollection} is passed to the
\code{x} argument).}
}
\value{
A \code{data.frame} counterpart object of the \code{by} argument's object.
The \code{data.frame} of the returned object contains an attribute 'count' representing the
number of counted Track objects for each spatial or spatio-temporal geometry of \code{by}.
}
\description{
Counting trajectories over spatial or spatio-temporal geometries. An \code{Track} object is
counted for a particualr geometry if at least one of its points
spatially or spatio-temporally intersects that particular geometry.

Works also with \code{data.frame} counterparts for the \code{by} argument.
}
\examples{
## Example 1 with toy data
## load example toy data
load(system.file("extdata","trajaggr_TestData.RData", package = "trajaggr"), verbose = FALSE)
countTrC_res <- count(TrColl, spPolys)
class(countTrC_res)
dim(spPolys)
dim(countTrC_res)
slot(countTrC_res, "data")
countTrC_byID <- count(TrColl, spPolys, byID = TRUE)
class(countTrC_byID)
head(slot(countTrC_byID, "data"))
countTrC_bySTF <- count(TrColl, stf_Polys_2t)
class(countTrC_bySTF)
slot(countTrC_bySTF, "data")
##
## Example 2 with wild boars example data
## load data
data(wildboars_4Ind_ltraj)
wb_TrColl <- as(wildboars_4Ind_ltraj, "TracksCollection")
library(sp)
wb_spPoly <- createSpatialArealObjFromPoints(as(wb_TrColl,
 "SpatialPointsDataFrame"), desDim = 8, out = "SpatialPixels")
wb_count <- count(wb_TrColl, wb_spPoly)
slot(wb_count, "data")
}

