% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createArealGeometryObjects.R
\name{createSpatialArealObjFromPoints}
\alias{createSpatialArealObjFromPoints}
\title{Function to create objects inheriting from class \code{Spatial} with an area-measured character}
\usage{
createSpatialArealObjFromPoints(x, desDim = 10, out = "SpatialPolygons")
}
\arguments{
\item{x}{an object of class \code{SpatialPoints} or \code{SpatialPointsDataFrame}}

\item{desDim}{integer indicating the desired dimension of the return object along
the larger side of the bbox of the input object \code{x}}

\item{out}{the desired class of the return object}
}
\value{
an object of class \code{SpatialPolygons} (default), \code{SpatialPixels} or
\code{SpatialGrid} as defined by the \code{out} argument
}
\description{
Here comes the description .... TODO
}
\examples{
## load data
## Load example toy data
load(system.file("extdata","trajaggr_TestData.RData", package = "trajaggr"), verbose = FALSE)
require(sp)
# Create SpatialPolygons covering points of a TracksCollection
points <- as(TrColl, "SpatialPointsDataFrame")
spPoly <- createSpatialArealObjFromPoints(points)
class(spPoly)
# Create SpatialPixels
spPix <- createSpatialArealObjFromPoints(points, desDim = 8, out = "SpatialPixels")
class(spPix)
spPix
# Create a SpatiaGrid
pointsB1 <- as(Track_B1, "SpatialPointsDataFrame")
spGrid_B1 <- createSpatialArealObjFromPoints(pointsB1, desDim = 9, out = "SpatialGrid")
class(spGrid_B1)
slot(spGrid_B1, "grid")
}

