% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_creation_and_documentation.R
\docType{data}
\name{pigeon_R_moveSt_sub}
\alias{pigeon_R_moveSt_sub}
\title{\code{pigeon_R_moveSt_sub} - small subset of the pigeons' example data from movebank}
\format{An object of class \code{MoveStack}}
\source{
Santos CD, Neupert S, Lipp H-P, Wikelski M, Dechmann D (2014) Temporal and contextual consistency of leadership in homing pigeon flocks. PLOS ONE 9(7): e102771.

Santos CD, Neupert S, Lipp H, Wikelski M, Dechmann D (2014) Data from: Temporal and contextual consistency of leadership in homing pigeon flocks. Movebank Data Repository.

Wikelski, M., and Kays, R. 2014. Movebank: archive, analysis and sharing of animal movement data. World Wide Web electronic publication. http://www.movebank.org, accessed on 2014-11-16.
}
\usage{
data(pigeon_R_moveSt_sub)
}
\description{
A small subset of the example data from movebank containing one object of
class \code{MoveStack}. This subset is used for illustration in the vignette.
}
\details{
The subset contains a \code{MoveStack} object containing two \code{Move} objects representing
pigeon trajectories of 15 seconds respectively 60 trajectory points. For further details
run \code{?pigeon_R_moveStack}.

The permission to use the data was obtained from the owners of the data (Contact person: Carlos David Santos).
}
\examples{
## load example data
## data(pigeon_R_moveSt_sub, package = "trajaggr")
data(pigeon_R_moveSt_sub)
class(pigeon_R_moveSt_sub)
# names of contained Move objects
levels(slot(pigeon_R_moveSt_sub, "trackId"))
head(slot(pigeon_R_moveSt_sub, "data"))
# for further details run
# ?pigeon_R_moveSt_sub
# ?pigeon_R_moveStack
}
\keyword{datasets}

