% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_creation_and_documentation.R
\docType{data}
\name{wildboars_4Ind_ltraj}
\alias{wildboars_4Ind_ltraj}
\title{\code{wildboars_4Ind_ltraj} - (adjusted) example data from the package \pkg{adehabititLT}}
\format{An object of class \code{ltraj}}
\source{
Original data source:

Maillard, D. (1996). Occupation et utilisation de la garrigue et du vignoble mediterraneens par
le Sanglier. Universite d'Aix-Marseille III: PhD thesis.

taken from \pkg{adehabititLT}:

Calenge, C. (2006) The package adehabitat for the R software: a tool for the analysis of space
and habitat use by animals. Ecological Modelling, 197, 516-519
}
\usage{
data(wildboars_4Ind_ltraj)
}
\description{
Five trajectories of four tracked wild boars. A modiefied data set from the package \pkg{adehabititLT}.
}
\details{
This data set is a modified version of the trajectory data stored in the object
\code{puechabonsp}, which is provided by the package \pkg{adehabititLT}. The data represent
the results of the monitoring of 4 wild boars in 1993 at
Puechabon (Mediterranean habitat, South of France) and thus the original data
contains four trajectories. The data set is modified in a way,
that the trajectory of one individual is split into two trajectories respectively bursts.
}
\examples{
## load example data
## data(wildboars_4Ind_ltraj, package = "trajaggr")
data(wildboars_4Ind_ltraj)
class(wildboars_4Ind_ltraj)
#if (requireNamespace("adehabitatLT", quietly = TRUE)) {
 adehabitatLT::summary.ltraj(wildboars_4Ind_ltraj)
#}
# for further details run
# ?wildboars_4Ind_ltraj
}
\keyword{datasets}

